// packages/server/src/lib/database.d.ts
declare const prisma: PrismaClient<
	{
		log: ("error" | "warn" | "query")[];
	},
	"error" | "warn" | "query",
	import("@prisma/client/runtime/client").DefaultArgs
>;
// ...
export declare const voiceService: {
	create(data: { username?: string; text: string; emotion: string; audioUrl?: string }): Promise<{
		text: string;
		id: string;
		username: string | null;
		createdAt: Date;
		emotion: string;
		audioUrl: string | null;
	}>;
	getRecent(limit?: number): Promise<
		{
			text: string;
			id: string;
			username: string | null;
			createdAt: Date;
			emotion: string;
			audioUrl: string | null;
		}[]
	>;
	getByUser(
		username: string,
		limit?: number,
	): Promise<
		{
			text: string;
			id: string;
			username: string | null;
			createdAt: Date;
			emotion: string;
			audioUrl: string | null;
		}[]
	>;
	deleteOldLogs(daysOld?: number): Promise<import("@prisma/client").Prisma.BatchPayload>;
};